requirements.txt
matplotlib==3.9.2
nltk==3.9.1
pandas==2.2.3
python-dotenv==1.0.1
tweepy==4.14.0
